effective_java/effective-java/src/test/java/me/whiteship/chapter04/item15/class_and_interfaces/item/ItemServiceTest.java
package me.whiteship.chapter04.item15.class_and_interfaces.item;

import me.whiteship.chapter04.item15.class_and_interfaces.member.MemberService;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

import static org.junit.jupiter.api.Assertions.*;

@ExtendWith(MockitoExtension.class)
class ItemServiceTest {

    @Mock
    MemberService memberService;

    @Test
    void itemService() {
        ItemService itemService = new ItemService(memberService);
        assertNotNull(itemService);
        assertNotNull(itemService.getMemberService());
//        assertNotNull(itemService.memberService);
    }
}